requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
azure-common==1.1.28
azure-core==1.38.0
azure-cosmos==4.14.6
azure-identity==1.25.1
azure-search-documents==11.6.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.4
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
eval-type-backport==0.1.3
exceptiongroup==1.3.1
Faker==37.12.0
fastapi==0.128.5
frozenlist==1.8.0
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lazy-object-proxy==1.12.0
MarkupSafe==3.0.3
more-itertools==10.8.0
motor==3.7.1
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.0.2
openai==2.17.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
parse==1.21.0
pathable==0.4.4
prance==23.6.21.0
propcache==0.4.1
pybars4==0.9.13
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyMeta3==0.5.1
pymongo==4.16.0
pyodbc==5.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.36.2
regex==2023.12.25
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.27.1
ruamel.yaml==0.19.1
scipy==1.13.1
semantic-kernel==0.9.6b1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.39.0
Werkzeug==3.1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
uvicorn
